Refactor FoldedFigure_Worker
